requirements.txt
streamlit
torch
torchvision
pillow
transformers